Rename TodoList examples (#543)
* Rename todolist to todolist-http-finch

* Move todolist-http-finch module to examples.todolist package

* Rename todolist-service example to todolist-lib
