+++ modules/examples/todolist-http-finch/src/main/scala/todo/http/apis/AppApi.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package todo
+package examples.todolist
 package http
 package apis
